server-py/pyproject.toml
[project]
name = "server-py"
version = "0.1.0"
description = "Python FastAPI backend scaffold for SKLinkChat"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.0,<2.0.0",
    "argon2-cffi>=25.1.0,<26.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.116.0,<1.0.0",
    "greenlet>=3.2.0,<4.0.0",
    "httpx>=0.28.0,<1.0.0",
    "pydantic-settings>=2.10.0,<3.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "redis>=6.4.0,<7.0.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.21.0,<1.0.0",
    "pytest>=8.4.0,<9.0.0",
    "pytest-asyncio>=1.1.0,<2.0.0",
    "ruff>=0.12.0,<1.0.0",
    "websockets>=15.0.0,<16.0.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
